+++ siga-ex/pom.xml
@@ -107,7 +107,7 @@
 			<artifactId>Tidy</artifactId>
 		</dependency>
 		<dependency>
-			<groupId>pdfbox</groupId>
+			<groupId>org.apache.pdfbox</groupId>
 			<artifactId>pdfbox</artifactId>
 		</dependency>
 		<!-- <dependency> <groupId>com.swetake.util</groupId> <artifactId>Qrcode</artifactId>
